# openapi_spec.yaml
openapi: 3.0.0
info:
  title: 티커 및 인디케이터 관리 API
  version: 1.0.0
  description: 티커 및 인디케이터를 관리하는 API입니다.

tags:
  - name: Ticker Management
    description: 티커 관리 관련 API
  - name: Indicator Management
    description: 인디케이터 관리 관련 API
  - name: Indicator Calculation
    description: 인디케이터 계산 관련 API

paths:
  /tickers:
    post:
      tags:
        - Ticker Management
      summary: 티커 추가
      description: 새로운 티커를 추가합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TickerInput'
      responses:
        '200':
          description: 성공적으로 티커가 추가되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerOutput'

    delete:
      tags:
        - Ticker Management
      summary: 티커 삭제
      description: 지정된 티커를 삭제합니다.
      parameters:
        - name: ticker
          in: query
          required: true
          description: 삭제할 티커 코드
          schema:
            type: string
            example: "AAPL"
      responses:
        '200':
          description: 성공적으로 티커가 삭제되었습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true

    put:
      tags:
        - Ticker Management
      summary: 티커 수정
      description: 기존 티커 정보를 수정합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TickerInput'
      responses:
        '200':
          description: 성공적으로 티커가 수정되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerOutput'

    get:
      tags:
        - Ticker Management
      summary: 티커 조회
      description: 지정된 티커 정보를 조회합니다.
      parameters:
        - name: ticker
          in: query
          required: true
          description: 조회할 티커 코드
          schema:
            type: string
            example: "AAPL"
      responses:
        '200':
          description: 성공적으로 티커 정보를 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerOutput'

  /tickers/list:
    get:
      tags:
        - Ticker Management
      summary: 티커 목록 조회
      description: 모든 티커 목록을 조회합니다.
      responses:
        '200':
          description: 성공적으로 티커 목록을 조회했습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickers:
                    type: array
                    items:
                      $ref: '#/components/schemas/TickerOutput'

  /indicators:
    post:
      tags:
        - Indicator Management
      summary: 인디케이터 추가
      description: 새로운 인디케이터를 추가합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndicatorInput'
      responses:
        '200':
          description: 성공적으로 인디케이터가 추가되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorOutput'

    delete:
      tags:
        - Indicator Management
      summary: 인디케이터 삭제
      description: 지정된 인디케이터를 삭제합니다.
      parameters:
        - name: indicatorCode
          in: query
          required: true
          description: 삭제할 인디케이터 코드
          schema:
            type: string
            example: "MACD"
      responses:
        '200':
          description: 성공적으로 인디케이터가 삭제되었습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true

    put:
      tags:
        - Indicator Management
      summary: 인디케이터 수정
      description: 기존 인디케이터 정보를 수정합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndicatorInput'
      responses:
        '200':
          description: 성공적으로 인디케이터가 수정되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorOutput'

    get:
      tags:
        - Indicator Management
      summary: 인디케이터 조회
      description: 지정된 인디케이터 정보를 조회합니다.
      parameters:
        - name: indicatorCode
          in: query
          required: true
          description: 조회할 인디케이터 코드
          schema:
            type: string
            example: "MACD"
      responses:
        '200':
          description: 성공적으로 인디케이터 정보를 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorOutput'

  /indicators/list:
    get:
      tags:
        - Indicator Management
      summary: 인디케이터 목록 조회
      description: 모든 인디케이터 목록을 조회합니다.
      responses:
        '200':
          description: 성공적으로 인디케이터 목록을 조회했습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  indicators:
                    type: array
                    items:
                      $ref: '#/components/schemas/IndicatorOutput'

  /calculate-indicator:
    post:
      tags:
        - Indicator Calculation
      summary: 인디케이터 계산
      description: 지정된 인디케이터를 계산합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                indicatorCode:
                  type: string
                  example: "MACD"
                ticker:
                  type: string
                  example: "AAPL"
                date:
                  type: string
                  format: date
                  example: "2023-01-01"
      responses:
        '200':
          description: 성공적으로 인디케이터가 계산되었습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
                  indicatorName:
                    type: string
                    example: "MACD"
                  indicatorCode:
                    type: string
                    example: "MACD"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2023-01-01T00:00:00Z"
                  modifiedAt:
                    type: string
                    format: date-time
                    example: "2023-01-01T00:00:00Z"

components:
  schemas:
    TickerInput:
      type: object
      properties:
        ticker:
          type: string
          example: "AAPL"
        name:
          type: string
          example: "Apple Inc."
        description:
          type: string
          example: "Apple Inc. 주식"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        modifiedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    TickerOutput:
      type: object
      properties:
        ticker:
          type: string
          example: "AAPL"
        tickerId:
          type: string
          example: "1234567890"
        name:
          type: string
          example: "Apple Inc."
        description:
          type: string
          example: "Apple Inc. 주식"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        modifiedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    IndicatorInput:
      type: object
      properties:
        indicatorName:
          type: string
          example: "MACD"
        indicatorCode:
          type: string
          example: "MACD"
        description:
          type: string
          example: "이동 평균 수렴 발산"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        modifiedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    IndicatorOutput:
      type: object
      properties:
        indicatorName:
          type: string
          example: "MACD"
        indicatorCode:
          type: string
          example: "MACD"
        description:
          type: string
          example: "이동 평균 수렴 발산"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        modifiedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"